Adds multiplayer functionality
After much stopping and learning along the way, I've use socket.io to send messages back and forth, utilizing them in the javascript (in server.js and app.js) to respond to them accordingly.

When a user connects they are alerted with a green square next to a Connected field. Once they place all their ships they can begin the game, triggering a cascade of side effects, the result in a Ready field also being updated.

If a user connects and readies before another even joins a check is done, and the incoming player will be notified.