Updates Single Player mode
The markup in singlePlayer.html has been refactored, removing buttons and relying on the URL to set the gameMode.

Code in the JS, app.js, had to also be refactored. Query selectors for the buttons have been removed, since they no longer exist or work to set the gameMode. The `ships` array and the calls to `createBoard` had to be hoisted in the code.