Fixes computer hit bug
Previously the computer was always getting a hit. I tracked down my problem to the `enemyGo` function. For one I was incorrectly referring the npc sunk ship scores (npcCarrierCount, npcDestroyerCount, etc, as cpuCarrierCount, cpuDestroyerCount etc) and I was not ever assigning a hit OR miss, just a hit.

Now that those have been solved for there are of course some additional bugs, but we'll get there!